pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["time9 <lol.time9@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^2.3.2"
gunicorn = "^20.1.0"
validators = "^0.20.0"
psycopg2-binary = "^2.9.6"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
beautifulsoup4 = "^4.12.2"

[tool.poetry.dev-dependencies]
pytest = "^7.0.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.flake8]
max-line-length = 88